Move naming trip patterns to after all transit data has been loaded
In order to only process them once
