+++ src/main/java/org/opentripplanner/graph_builder/GraphBuilder.java
@@ -24,6 +24,7 @@ import org.opentripplanner.graph_builder.module.OsmBoardingLocationsModule;
 import org.opentripplanner.graph_builder.module.PruneNoThruIslands;
 import org.opentripplanner.graph_builder.module.StreetLinkerModule;
 import org.opentripplanner.graph_builder.module.TimeZoneAdjusterModule;
+import org.opentripplanner.graph_builder.module.TripPatternNamer;
 import org.opentripplanner.graph_builder.module.map.BusRouteStreetMatcher;
 import org.opentripplanner.graph_builder.module.ned.DegreeGridNEDTileSource;
 import org.opentripplanner.graph_builder.module.ned.ElevationModule;
@@ -144,6 +145,10 @@ public class GraphBuilder implements Runnable {
 
     if (hasNetex) {
       graphBuilder.addModule(netexModule(config, dataSources.get(NETEX)));
+    }
+
+    if (hasTransitData) {
+      graphBuilder.addModule(new TripPatternNamer());
     }
 
     if (hasTransitData && graphBuilder.transitModel.getAgencyTimeZones().size() > 1) {
