+++ src/main/java/org/opentripplanner/graph_builder/module/AddTransitModelEntitiesToGraph.java
@@ -408,9 +408,6 @@ public class AddTransitModelEntitiesToGraph {
   private void addTripPatternsToTransitModel(TransitModel transitModel) {
     Collection<TripPattern> tripPatterns = otpTransitService.getTripPatterns();
 
-    /* Generate unique human-readable names for all the TableTripPatterns. */
-    TripPattern.generateUniqueNames(tripPatterns);
-
     /* Loop over all new TripPatterns setting the service codes. */
     for (TripPattern tripPattern : tripPatterns) {
       tripPattern.setServiceCodes(transitModel.getServiceCodes()); // TODO this could be more elegant
